Day 1
First part is based on just checking the past integer and the next.
Second part is based on cycling through input from user and appending it to a list, from where our integers will be summed and compared. Took me almost three hours to solve, but proud of my solution.